[refactor] 구조 변경으로 인한 경로 오류 수정 (#76)
what : 구조 변경으로 인한 경로 오류 수정 (#76)
why : 구조 변경으로 인한 경로 오류 수정 (#76)
