+++ Backend/src/main/java/com/github/backend/web/course/controller/CourseController.java
@@ -1,8 +1,10 @@
-package com.github.backend.web;
+package com.github.backend.web.course.controller;
 
-import com.github.backend.dto.course.CourseInfoResponse;
+import com.github.backend.dto.common.Result;
+import com.github.backend.dto.course.CourseInfoOutputDto;
 import com.github.backend.service.course.impl.CourseService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -15,12 +17,17 @@ public class CourseController {
     private final CourseService courseService;
 
     @GetMapping("/{courseId}")
-    public CourseInfoResponse getCourseInfo(
+    public ResponseEntity<Result<?>> getCourseInfo(
             @PathVariable("courseId") @Valid Long courseId
     ){
-        return CourseInfoResponse.from(
-                courseService.getCourseInfo(courseId)
-        );
+        return ResponseEntity.ok()
+                .body(
+                        Result.builder()
+                                .status(200)
+                                .success(true)
+                                .data(courseService.getCourseInfo(courseId))
+                                .build()
+                );
     }
