+++ Backend/src/main/java/com/github/backend/web/course/controller/GamerController.java
@@ -1,4 +1,4 @@
-package com.github.backend.web;
+package com.github.backend.web.course.controller;
 
 import com.github.backend.dto.common.Result;
 import com.github.backend.dto.gamer.SearchGamer;
